requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
beautifulsoup4>=4.11.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
ipykernel>=6.15.0
tqdm>=4.64.0
pyyaml>=6.0
jsonlines>=3.0.0

# Model Training Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
accelerate>=0.20.0
evaluate>=0.4.0
wandb>=0.15.0

# Graph Neural Networks
torch-geometric>=2.3.0
networkx>=3.1.0
dgl>=1.1.0

# Additional ML Libraries
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.2.0
shap>=0.42.0

# Visualization
plotly>=5.15.0
bokeh>=3.2.0

# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0

# API Testing and Load Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
locust>=2.17.0
pytest-cov>=4.1.0

# API Documentation
swagger-ui-bundle>=0.1.0
redoc>=0.1.0

# Database and Caching
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Task 5: Security Validation and Deployment Dependencies
# Static Analysis Tools
slither-analyzer>=0.10.0
mythril>=0.24.0
crytic-compile>=0.3.0

# QIE SDK and Blockchain Integration
qie-sdk>=1.0.0
web3>=6.11.0
eth-account>=0.9.0
eth-utils>=2.3.0

# Deployment and Containerization
docker>=6.1.0
docker-compose>=1.29.0
kubernetes>=28.1.0

# Performance Testing and Benchmarking
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0
psutil>=5.9.0

# Documentation Generation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Additional Testing Tools
pytest-xdist>=3.3.0
coverage>=7.3.0
